Fix startup crash caused by inconsistency in the saving and parsing of dates
* Fix startup crash caused by inconsistency in the saving and parsing of dates

The "Last played" date is saved using a non-changing locale, which stays the same
on all systems. However, the date is parsed with no specific locale, causing the
parsing method to default to the system locale.

The expected locale uses a month-day-year format, but if the player's system
locale uses a different format, such as day-month-year, certain dates will look
invalid and trigger a crash, while others will just be parsed incorrectly.

Formats other than month-day-year are the default in most of the world outside
of the United States, so this bug likely affects a significant amount of the
playerbase.

Here are two examples, where the system uses day-month-year:
* July 26th, 2024 gets saved as "07-26-2024", but there's no 26th month, so it
  crashes.
* January 8th, 2024 gets saved as "01-08-2024", which gets parsed as August 1st,
  2024.

Signed-off-by: Ashe <hashedpanda@outlook.dk>